completing improvements in ncurses_control module

+++ modules/live/ncurses_control.h
@@ -21,7 +21,7 @@ void println_to_window(int window_handle, char *message);
 void print_int_to_window(int window_handle, int value);
 void print_double_to_window(int window_handle, double value);
 void print_intmsg_to_window(int window_handle, char *message, int value);
-void print_doublemsg_to_window(int window_handle, char *message, int value);
+void print_doublemsg_to_window(int window_handle, char *message, double value);
 
 void add_key_listener(char *context, void (callback(int)));
 void remove_key_listener(char *context);
